refactor(terminal): split middleware boundaries
